feat : 발표전 table 수정완료
