+++ src/pages/admin/Business/RoomPage/RoomListPage.js
@@ -1,4 +1,4 @@
-import { Checkbox, Table } from "antd";
+import { Table } from "antd";
 import React, { useEffect, useState } from "react";
 import {
   DogInfoButton,
@@ -24,42 +24,242 @@ import {
   RmTodaySearch,
   StyledTableWrap,
 } from "../../../../styles/AdminPageStyle/RoomPageStyle/roomPageStyle";
-import { getRoomList } from "../../../../api/admin/Room/RoomApi";
 
-const RoomListPage = () => {
+const RoomListTest = () => {
   const [selectedRowKeys, setSelectedRowKeys] = useState([]); // 선택된 행의 키 값들을 저장하는 상태
   const [selectedRows, setSelectedRows] = useState([]); // 선택된 행들의 정보를 저장하는 상태
-
-  // 초기값 설정해보기
-  const [initData, setInitData] = useState([]);
-  /* -------------------------- 체크박스  start ------------------------ */
-  // 각 행의 선택 여부를 저장하는 상태
-  const [rowSelections, setRowSelections] = useState({});
+  const [showAllData, setShowAllData] = useState(false); // 전체 목록을 보여줄지 여부를 저장하는 상태
+  const [cancelOpen, setCancelOpen] = useState(false); // 예약 취소 모달 오픈 여부 관리
 
   // 체크박스가 변경될 때 호출되는 함수
-  const handleSelectionChange = (Keys, Rows) => {
-    if (Keys.length > 1) {
-      // 선택된 행이 하나 이상인 경우, 첫 번째 행의 키 값만 선택하도록 설정
-      setSelectedRowKeys([selectedRowKeys[0]]);
-      setSelectedRows([selectedRows[0]]);
-      console.log("1개의 행만 선택되니?");
-    } else {
-      // 선택된 행이 하나인 경우, 선택된 행의 키 값과 정보를 업데이트
-      setSelectedRowKeys(selectedRowKeys);
-      setSelectedRows(selectedRows);
-      console.log("1개의 행만 선택되니?");
-    }
+  const handleSelectionChange = (keys, rows) => {
+    setSelectedRowKeys(keys); // 선택된 행의 키 값들을 업데이트
+    setSelectedRows(rows); // 선택된 행들의 정보를 업데이트
   };
+  /* 🙂 ant design table 적용해보자 */
+  // type RmReserve = {
+  //   checkbox: string, // 체크 : checkbox
+  //   number: string, // 번호 : number
+  //   reserveNumber: number, // 예약번호 : reserveNumber
+  //   nickname: number, // 닉네임 : nickname
+  //   roomType: string, // 객실유형 : roomType
+  //   dogInfo: number, // 반려견정보 : dogInfo
+  //   reservationData: number, // 예약날짜(체크인아웃) : reservationData
+  //   phoneNumber: number, // 전화번호 : phoneNumber
+  //   paymentAmount: number, // 결제금액 : paymentAmount
+  //   status: number, // 상태 : status
+  // };
 
-  const rowSelection = {
-    onChange: (selectedRowKeys, selectedRows) => {
-      console.log(
-        `selectedRowKeys: ${selectedRowKeys}`,
-        "selectedRows: ",
-        selectedRows,
-      );
+  // 초기값 설정해보기
+  const [initData, setInitData] = useState([
+    {
+      checkbox: 1,
+      key: 1,
+      reserveNumber: 66666,
+      nickname: "누룽지",
+      roomType: "소형견(3kg ~7kg)이하 기준",
+      dogInfo: "반려견정보",
+      dogname: "룽지",
+      dogage: "4",
+      dogsize: "소형견",
+      reservationData: "20240101-20240103",
+      phoneNumber: "010-3333-5555",
+      paymentAmount: 45000,
+      status: "예약취소",
     },
-  };
+    {
+      checkbox: 1,
+      key: 2,
+      reserveNumber: 81945,
+      nickname: "콩지",
+      roomType: "중형견(7kg ~15kg)이하 기준",
+      dogInfo: "반려견정보",
+      dogname: "콩지",
+      dogage: "7",
+      dogsize: "중형견",
+      reservationData: "20240301-20240303",
+      phoneNumber: "010-2222-7777",
+      paymentAmount: 68000,
+      status: "이용완료",
+    },
+    {
+      checkbox: 1,
+      key: 3,
+      reserveNumber: 81945,
+      nickname: "콩지",
+      roomType: "중형견(7kg ~15kg)이하 기준",
+      dogInfo: "반려견정보",
+      dogname: "콩지마미",
+      dogage: "10",
+      dogsize: "중형견",
+      reservationData: "20240301-20240303",
+      phoneNumber: "010-2222-7777",
+      paymentAmount: 68000,
+      status: "이용완료",
+    },
+    {
+      checkbox: 1,
+      key: 4,
+      reserveNumber: 81945,
+      nickname: "콩지",
+      roomType: "중형견(7kg ~15kg)이하 기준",
+      dogInfo: "반려견정보",
+      dogname: "콩지대디",
+      dogage: "10",
+      dogsize: "중형견",
+      reservationData: "20240301-20240303",
+      phoneNumber: "010-2222-7777",
+      paymentAmount: 68000,
+      status: "이용완료",
+    },
+    {
+      checkbox: 1,
+      key: 5,
+      reserveNumber: 718382,
+      nickname: "나강아지키운다",
+      roomType: "소형견(7kg ~)이하 기준",
+      dogInfo: "반려견정보",
+      dogname: "미미",
+      dogage: "2",
+      dogsize: "소형견",
+      reservationData: "20240308-20240310",
+      phoneNumber: "010-1412-9321",
+      paymentAmount: 50000,
+      status: "입실완료",
+    },
+    {
+      checkbox: 1,
+      key: 6,
+      reserveNumber: 718382,
+      nickname: "나강아지키운다",
+      roomType: "중형견(7kg ~15kg)이하 기준",
+      dogInfo: "반려견정보",
+      dogname: "포포",
+      dogage: "5",
+      dogsize: "소형견",
+      reservationData: "20240308-20240310",
+      phoneNumber: "010-1412-9321",
+      paymentAmount: 50000,
+      status: "입실완료",
+    },
+    {
+      checkbox: 1,
+      key: 7,
+      reserveNumber: 111423,
+      nickname: "랩탈리언",
+      roomType: "중형견(7kg ~15kg)이하 기준",
+      dogInfo: "반려견정보",
+      dogname: "포포",
+      dogage: "5",
+      dogsize: "소형견",
+      reservationData: "20240310-20240311",
+      phoneNumber: "010-5134-6281",
+      paymentAmount: 72000,
+      status: "예약취소",
+    },
+    {
+      checkbox: 1,
+      key: 8,
+      reserveNumber: 81945,
+      nickname: "바다의왕자",
+      roomType: "대형견(15kg ~40kg)이하 기준",
+      dogInfo: "반려견정보",
+      dogname: "포포",
+      dogage: "5",
+      dogsize: "소형견",
+      reservationData: "20240320-20240322",
+      phoneNumber: "010-8415-6319",
+      paymentAmount: 49000,
+      status: "예약대기중",
+    },
+    {
+      checkbox: 1,
+      key: 9,
+      reserveNumber: 81945,
+      nickname: "머피",
+      roomType: "소형견(7kg ~)이하 기준",
+      dogInfo: "반려견정보",
+      dogname: "포포",
+      dogage: "5",
+      dogsize: "소형견",
+      reservationData: "20240314-20240315",
+      phoneNumber: "010-2496-8195",
+      paymentAmount: 77000,
+      status: "예약취소",
+    },
+    {
+      checkbox: 1,
+      key: 10,
+      reserveNumber: 0,
+      nickname: "",
+      roomType: "중형견(7kg ~15kg)이하 기준",
+      dogInfo: "",
+      reservationData: "",
+      phoneNumber: "",
+      paymentAmount: 0,
+      status: "예약없음",
+    },
+    {
+      checkbox: 1,
+      key: 11,
+      reserveNumber: 3022961,
+      nickname: "유저11",
+      roomType: "대형견(15kg ~40kg)이하 기준",
+      dogInfo: "반려견정보",
+      dogname: "풍랑",
+      dogage: "4",
+      dogsize: "대형견",
+      reservationData: "20240314-20240320",
+      phoneNumber: "01046858912",
+      paymentAmount: 248000,
+      status: "예약대기중",
+    },
+    {
+      checkbox: 1,
+      key: 12,
+      reserveNumber: 3022961,
+      nickname: "유저11",
+      roomType: "대형견(15kg ~40kg)이하 기준",
+      dogInfo: "반려견정보",
+      dogname: "파파",
+      dogage: "7",
+      dogsize: "대형견",
+      reservationData: "20240314-20240320",
+      phoneNumber: "01046858912",
+      paymentAmount: 248000,
+      status: "예약대기중",
+    },
+    {
+      checkbox: 1,
+      key: 13,
+      reserveNumber: 3022961,
+      nickname: "유저11",
+      roomType: "대형견(15kg ~40kg)이하 기준",
+      dogInfo: "반려견정보",
+      dogname: "자갈치",
+      dogage: "2",
+      dogsize: "대형견",
+      reservationData: "20240314-20240320",
+      phoneNumber: "01046858912",
+      paymentAmount: 248000,
+      status: "예약대기중",
+    },
+    {
+      checkbox: 1,
+      key: 14,
+      reserveNumber: 3022961,
+      nickname: "유저11",
+      roomType: "대형견(15kg ~40kg)이하 기준",
+      dogInfo: "반려견정보",
+      dogname: "뭉탱",
+      dogage: "6",
+      dogsize: "대형견",
+      reservationData: "20240314-20240320",
+      phoneNumber: "01046858912",
+      paymentAmount: 248000,
+      status: "예약대기중",
+    },
+  ]);
   /* -------------------------- 필터 start ------------------------ */
   /* 💚💚💚예약번호, 닉네임 검색 -> 해당 row 가 뜨도록하는 함수 */
   const [searchText, setSearchText] = useState(""); // 검색어 상태 관리
@@ -73,7 +273,8 @@ const RoomListPage = () => {
     // 검색어에 해당하는 행 필터링하여 새로운 데이터 생성
     const filtered = initData.filter(
       item =>
-        item.resNum.toString().includes(value) || item.nickname.includes(value),
+        item.reserveNumber.toString().includes(value) ||
+        item.nickname.includes(value),
     );
     setFilteredData(filtered); // 필터링된 데이터 업데이트
   };
@@ -92,29 +293,18 @@ const RoomListPage = () => {
   });
   /* -------------------------- 필터 end ------------------------ */
   const columns = [
-    // {
-    //   title: "✔", // 컬럼 제목을 "Checkbox"로 변경
-    //   key: "selection",
-    //   render: (text, record) => (
-    //     <Checkbox
-    //       checked={rowSelections[record.resPk]} // 해당 행의 선택 여부를 체크박스에 반영합니다.
-    //       onChange={() => handleCheckboxChange(record)} // 체크박스 변경 이벤트 핸들러를 등록합니다.
-    //     />
-    //   ),
-    // },
     {
       title: "번호",
-      dataIndex: "resPk",
+      dataIndex: "key",
       // defaultSortOrder: "descend",
       sorter: (a, b) => a.number - b.number,
-      defaultSortOrder: null, // 이 부분을 추가하여 기본 정렬을 해제합니다.
-      key: "resPk",
+      key: "key",
     },
     {
       title: "예약번호",
-      dataIndex: "resNum",
-      key: "resNum",
-      ...getColumnSearchProps("resNum"),
+      dataIndex: "reserveNumber",
+      key: "reserveNumber",
+      ...getColumnSearchProps("reserveNumber"),
     },
     {
       title: "닉네임",
@@ -124,55 +314,35 @@ const RoomListPage = () => {
     },
     {
       title: "객실유형",
-      dataIndex: "hotelRoomNm",
-      key: "hotelRoomNm",
+      dataIndex: "roomType",
+      key: "roomType",
     },
     {
       title: "반려견정보",
-      dataIndex: "resDogPk",
-      key: "resDogPk",
-      // 반려견정보 버튼 클릭 시, 모달 창 뜨도록 하게!
-      render: (text, row) => {
-        if (!text) {
-          return null; // resDogPk 없을 때 빈 값 반환
-        }
-        return (
-          <DogInfoButton onClick={() => dogInfoModalOpen(text)}>
-            {"반려견정보"}
+      dataIndex: "dogInfo",
+      key: "dogInfo",
+      render: (text, record) => {
+        return text ? (
+          <DogInfoButton onClick={() => openDogInfoModal(record)}>
+            {text}
           </DogInfoButton>
-        );
+        ) : null;
       },
     },
     {
       title: "예약날짜(체크인아웃)",
       dataIndex: "reservationData",
       key: "reservationData",
-      render: (text, row) => (
-        <span>
-          {row.fromDate} - {row.toDate}
-        </span>
-      ),
     },
     {
       title: "전화번호",
-      dataIndex: "userPhoneNum",
-      key: "userPhoneNum",
-      render: (text, row) => {
-        if (!text) {
-          return null; // row 또는 userPhoneNum이 없으면 빈 값 반환
-        }
-        // 전화번호를 '-'로 구분하여 세 부분으로 나누고 합치는 방법
-        const formattedPhoneNumber = `${row.userPhoneNum.slice(
-          0,
-          3,
-        )}-${row.userPhoneNum.slice(3, 7)}-${row.userPhoneNum.slice(7)}`;
-        return <span>{formattedPhoneNumber}</span>;
-      },
+      dataIndex: "phoneNumber",
+      key: "phoneNumber",
     },
     {
       title: "결제금액",
-      dataIndex: "payment",
-      key: "payment",
+      dataIndex: "paymentAmount",
+      key: "paymentAmount",
       render: text => {
         if (!text) {
           return null; // 값이 없을 경우 빈 값 반환
@@ -184,26 +354,34 @@ const RoomListPage = () => {
     },
     {
       title: "상태",
-      key: "resStatus",
-      dataIndex: "resStatus",
+      key: "status",
+      dataIndex: "status",
+      /* 
+      render 옵션은 Array.map()처럼 작동합니다.
+      render: (text, row, index) => {};
+      text: name의 data [String]
+      row: 하나의 row data [Object]
+      index: row index [Number]
+      */
+      // 예약취소일때는 button 추가하여 모달 뜨도록 하도록 render
       render: (text, row) => {
-        if (row.resStatus === 0) {
-          return "예약대기중";
-        } else if (row.resStatus === 1) {
-          return "예약완료";
-        } else if (row.resStatus === 2) {
-          return "이용중";
-        } else if (row.resStatus === 3) {
-          return "이용완료";
-        } else if (row.resStatus === 4 || row.resStatus === 5) {
-          return (
-            <DogInfoButton onClick={() => cancelModalOpen(text)}>
-              {"예약취소"}
-            </DogInfoButton>
-          );
-        } else {
-          return null;
-        }
+        return text === "예약취소" ? (
+          <DogInfoButton onClick={() => cancelModalOpen(text)}>
+            {text}
+          </DogInfoButton>
+        ) : text === "예약완료" ? (
+          "예약완료"
+        ) : text === "입실완료" ? (
+          "입실완료"
+        ) : text === "퇴실완료" ? (
+          "퇴실완료"
+        ) : text === "예약없음" ? (
+          "예약없음"
+        ) : text === "예약대기중" ? (
+          "예약대기중"
+        ) : text === "이용완료" ? (
+          "이용완료"
+        ) : null;
       },
       // 💥💥💥💥💥💥 변경요함
       // 행 필터 추가
@@ -235,28 +413,19 @@ const RoomListPage = () => {
     },
   ];
 
-  // // 선택된 행의 정보 저장
-  // const rowSelection = {
-  //   // 선택된 행의 키 값, 선택된행의 정보를 매개변수로 받아
-  //   onChange: (selectedRowKeys, selectedRows) => {
-  //     console.log(
-  //       `selectedRowKeys: ${selectedRowKeys}`,
-  //       "selectedRows: ",
-  //       selectedRows,
-  //     );
-  //   },
-  // };
+  // 선택된 행의 정보 저장
+  const rowSelection = {
+    // 선택된 행의 키 값, 선택된행의 정보를 매개변수로 받아
+    onChange: (selectedRowKeys, selectedRows) => {
+      console.log(
+        `selectedRowKeys: ${selectedRowKeys}`,
+        "selectedRows: ",
+        selectedRows,
+      );
+    },
+  };
 
-  // // 체크박스가 변경될 때 호출되는 함수
-  // const handleSelectionChange = (selectedRowKeys, selectedRows) => {
-  //   setSelectedRowKeys(selectedRowKeys); // 선택된 행의 키 값들을 업데이트
-  //   setSelectedRows(selectedRows); // 선택된 행들의 정보를 업데이트
-  //   console.log("체크박스에 선택되냐?");
-
-  //   // 위의 두 줄을 추가하여 선택된 행의 정보를 업데이트합니다.
-  // };
-
-  // 체크->(예약승인) 버튼 클릭 시, 상태 변경 함수
+  // 체크->(예약승인)버튼 클릭 시, 상태 변경 함수
   const handleReservationAp = (selectedRowKeys, selectedRows) => {
     // 체크박스가 선택되지 않았을 때 알림창 표시
     if (selectedRowKeys.length === 0) {
@@ -277,82 +446,70 @@ const RoomListPage = () => {
     setInitData(updatedData);
     console.log("Updated data:", updatedData);
   };
-  const handleCheckInCom = () => {};
-  const handleCheckOutCom = () => {};
-
-  /* ---------------------- 💛 axios 연동 start --------------------*/
-  /* room today 초기값
-  dogSizeNm: "소형견"
-  dogSizePk: 1
-  fromDate: "2024-03-05"
-  hotelRoomNm: "소형견(7kg 이하)"
-  hotelRoomPk: 1
-  nickname: "백서윤"
-  paymentAmount: 150000
-  resDogAge: 4
-  resDogInfo: "안물어요"
-  resDogNm: "뽀송이"
-  resDogPk: 1
-  resNum: "R2435874146183"
-  resPk: 1
-  resStatus: 0
-  toDate: "2024-03-07"
-  userPhoneNum: "01023885447"
-
-  fromDate: "2024-03-05"
-  hotelNm: "멍뭉이 스타일 애견 호텔"
-  nickname: "백서윤"
-  payment: 150000
-  resNum: "R2435874146183"
-  resPk: 1
-  resStatus: 2
-  toDate: "2024-03-07"
-  userPhoneNum: "01023885447"
-  */
-  // 첫페이지는 1이므로 초기값을 1로 설정
-  const [page, setPage] = useState(1);
-  useEffect(() => {
-    // 페이지 처음 불러들일 때 실행
-    getRoomList(page, successGetRoomList, failGetRoomList, errorGetRoomList);
-    console.log("현재 페이지: ", page);
-    //page의 값이 바뀌면 useEffect를 다시 실행하겠다.
-  }, [page]);
-
-  const successGetRoomList = result => {
-    console.log("성공했습니다.", result);
-    setInitData(result.reservationInfoList); // 서버에서 받은 데이터를 설정
+  const handleCheckInCom = (selectedRowKeys, selectedRows) => {
+    // 체크박스가 선택되지 않았을 때 알림창 표시
+    if (selectedRowKeys.length === 0) {
+      alert("예약할 항목을 선택해주세요.");
+      return; // 함수 종료
+    }
+    console.log("selectedRowKeys:", selectedRowKeys);
+    console.log("selectedRows:", selectedRows);
+    // 기존의 데이터를 복제하여 새로운 배열 생성
+    const updatedData = [...initData];
+    // 업데이트된 데이터를 새로운 배열에 추가
+    updatedData.forEach(row => {
+      if (selectedRowKeys.includes(row.key) && row.status !== "이용완료") {
+        row.status = "이용완료";
+      }
+    });
+    // 새로운 배열로 initData 업데이트
+    setInitData(updatedData);
+    console.log("Updated data:", updatedData);
   };
-  const failGetRoomList = result => {
-    console.log("다시 시도해주세요.", result);
+  const handleCheckOutCom = (selectedRowKeys, selectedRows) => {
+    // 체크박스가 선택되지 않았을 때 알림창 표시
+    if (selectedRowKeys.length === 0) {
+      alert("예약할 항목을 선택해주세요.");
+      return; // 함수 종료
+    }
+    console.log("selectedRowKeys:", selectedRowKeys);
+    console.log("selectedRows:", selectedRows);
+    // 기존의 데이터를 복제하여 새로운 배열 생성
+    const updatedData = [...initData];
+    // 업데이트된 데이터를 새로운 배열에 추가
+    updatedData.forEach(row => {
+      if (selectedRowKeys.includes(row.key) && row.status !== "퇴실완료") {
+        row.status = "퇴실완료";
+      }
+    });
+    // 새로운 배열로 initData 업데이트
+    setInitData(updatedData);
+    console.log("Updated data:", updatedData);
   };
-  const errorGetRoomList = result => {
-    console.log("서버 에러입니다.", result);
-  };
-
-  /* ---------------------- 💛 axios 연동 end --------------------*/
 
   /* ----------------------💭 모달 open & close start  --------------------*/
-  const [dogInfoOpen, setDogInfoOpen] = useState(false); // 반려견 정보 모달 오픈 여부 관리
-  const [cancelOpen, setCancelOpen] = useState(false); // 예약 취소 모달 오픈 여부 관리
-  // 💥 반려견 정보 모달 관련 내용 추가!
-  // RoomPage 컴포넌트에서 선택된 행 정보를 저장할 상태 추가
+  // 예약 취소 및 반려견정보 선택 시, 모달 오픈
+
+  const [dogInfoOpen, setDogInfoOpen] = useState(false);
   const [selectedDogInfo, setSelectedDogInfo] = useState(null);
 
-  const cancelModalOpen = row => {
-    // setSelectedRow(row); // 선택된 행 정보 저장0
-    setCancelOpen(true); // 예약 취소 모달 오픈
-    setDogInfoOpen(false); // 반려견 정보 모달 닫기
+  const cancelModalOpen = text => {
+    // 여기에 조건부 논리 추가
+    if (text === "예약취소") {
+      setCancelOpen(true);
+      setDogInfoOpen(false); // 반려견 정보 모달 닫기 ?????
+    }
   };
-  const dogInfoModalOpen = row => {
-    // setSelectedRow(row); // 선택된 행 정보 저장
-    setSelectedDogInfo(row); // 선택된 행의 반려견 정보 저장
-    setDogInfoOpen(true); // 반려견 정보 모달 오픈
-    setCancelOpen(false); // 예약 취소 모달 닫기
+  // 반려견 정보 모달 열기
+  const openDogInfoModal = dogInfo => {
+    setSelectedDogInfo(dogInfo);
+    setDogInfoOpen(true);
   };
   // 모달 close 버튼 클릭시, 닫도록
-  const cancelModalClose = () => {
-    setCancelOpen(false);
-    setDogInfoOpen(false);
+  const cancelModalClose = input => {
+    console.log("콘솔은 닫힌다.");
+    setCancelOpen(false); // 취소 모달 닫기
+    setDogInfoOpen(false); // 반려견 정보 모달 닫기
   };
   /* ----------------------💭 모달 open & close end --------------------*/
 
@@ -361,6 +518,29 @@ const RoomListPage = () => {
       {/* header 영역 */}
       <div>
         <RmMenuSearchFlex>
+          {/* <RmTodayMenu>
+            <RmTodayMenuBt>전체</RmTodayMenuBt>
+            <img
+              src={`${process.env.PUBLIC_URL}/admin/images/RmToday/bar.svg`}
+              alt=""
+            />
+            <RmTodayMenuBt>이용중</RmTodayMenuBt>
+            <img
+              src={`${process.env.PUBLIC_URL}/admin/images/RmToday/bar.svg`}
+              alt=""
+            />
+            <RmTodayMenuBt>예약완료</RmTodayMenuBt>
+            <img
+              src={`${process.env.PUBLIC_URL}/admin/images/RmToday/bar.svg`}
+              alt=""
+            />
+            <RmTodayMenuBt>대기중</RmTodayMenuBt>
+            <img
+              src={`${process.env.PUBLIC_URL}/admin/images/RmToday/bar.svg`}
+              alt=""
+            />
+            <RmTodayMenuBt>이용완료</RmTodayMenuBt>
+          </RmTodayMenu> */}
           <div>
             <RmTodaySearch
               type="text"
@@ -380,24 +560,29 @@ const RoomListPage = () => {
             >
               예약완료
             </RmPageBt>
-            <RmPageBt onClick={() => handleCheckInCom()}>입실완료</RmPageBt>
-            <RmPageBt onClick={() => handleCheckOutCom()}>퇴실완료</RmPageBt>
+            <RmPageBt
+              onClick={() => handleCheckInCom(selectedRowKeys, selectedRows)}
+            >
+              입실완료
+            </RmPageBt>
+            <RmPageBt
+              onClick={() => handleCheckOutCom(selectedRowKeys, selectedRows)}
+            >
+              퇴실완료
+            </RmPageBt>
           </RmBtFlex>
           <StyledTableWrap>
             <Table
-              // rowSelection={rowSelection}
+              // 행 선택
               rowSelection={{
-                type: "checkbox", // 하나의 체크박스만 선택할 수 있도록 설정
-                onChange: handleSelectionChange, // handleSelectionChange 함수로 변경
+                type: "checkbox",
+                // 체크박스가 변경될 때 호출되는 콜백 함수
+                onChange: handleSelectionChange,
               }}
-              dataSource={
-                filteredData.length > 0
-                  ? filteredData.map(item => ({ ...item, key: item.resPk }))
-                  : initData.map(item => ({ ...item, key: item.resPk }))
-              }
-              // dataSource={filteredData.length > 0 ? filteredData : initData}
+              dataSource={filteredData}
               columns={columns}
               pagination={{
+                // 페이지 네이션
                 pageSize: 10,
                 position: ["bottomCenter"],
                 hideOnSinglePage: false,
@@ -465,12 +650,14 @@ const RoomListPage = () => {
         </ModalBackground>
       )}
       {/* 반려견 정보 모달 */}
-      {dogInfoOpen && (
+      {/* {cancelModalOpen && (
+        <RmPageModal cancelModalClose={cancelModalClose(setCancelOpen)} />
+      )} */}
+      {dogInfoOpen && selectedDogInfo && (
         <ModalBackground>
           <RmPageModal>
             <RmModalClose
-              // 모달 창 닫을때 함수 실행
-              onClick={cancelModalClose}
+              onClick={() => setDogInfoOpen(false)}
               src={`${process.env.PUBLIC_URL}/admin/images/RmToday/close.svg`}
               alt=""
             />
@@ -480,17 +667,13 @@ const RoomListPage = () => {
             <RmModalDogContent>
               <RmModalDogHead>
                 <RmModalDogTitle>
-                  객실유형 : {selectedDogInfo.hotelRoomNm}
+                  객실유형 : {selectedDogInfo.roomType}
                 </RmModalDogTitle>
               </RmModalDogHead>
-              {/* <img
-                src={`${process.env.PUBLIC_URL}/admin/images/RmToday/exampleimg.svg`}
-                alt=""
-              /> */}
               <RmDogInfo>
-                <span>강아지 이름 : {selectedDogInfo.resDogNm}</span>
-                <span>강아지 나이 : {selectedDogInfo.resDogAge}</span>
-                <span>강아지 크기 : {selectedDogInfo.dogSizeNm}</span>
+                <span>강아지 이름 : {selectedDogInfo.dogname}</span>
+                <span>강아지 나이 : {selectedDogInfo.dogage}</span>
+                <span>강아지 크기 : {selectedDogInfo.dogsize}</span>
               </RmDogInfo>
             </RmModalDogContent>
           </RmPageModal>
@@ -500,4 +683,4 @@ const RoomListPage = () => {
   );
 };
 
-export default RoomListPage;
+export default RoomListTest;
